requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
nose==1.3.7
pathlib2==2.3.2
pbr==5.1.1
pluggy==0.8.0
py==1.7.0
pylint==2.2.0
pymediawiki==0.5.0
pytest==4.0.0
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wikipedia==1.4.0
wrapt==1.10.11
